Moved logic to TweetFeedActivity & added notifications
